Finished Initialization
I updated the initialization function, most of it needed to be changed.
I deleted the ADC initialization because it is different for this PIC
and some registers such as ADCCON0 will need to be changed in each
function in order to use different ADC ports. Before, we only used one
ADC port so it was static. The port for the battery and throttle buffer
circuit was changed from RA6,RA7, to RB0 and RB1. Previous ports had no
analog channel. New schematic was emailed out.